% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.qgarch}
\alias{fitted.qgarch}
\title{Fitted values from a qgarch model}
\usage{
\method{fitted}{qgarch}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qgarch} object.}

\item{...}{Unused.}
}
\value{
A numeric vector of fitted values.
}
\description{
Fitted values from a qgarch model
}
