% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.qgarch}
\alias{logLik.qgarch}
\title{Log-likelihood for a qgarch model}
\usage{
\method{logLik}{qgarch}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qgarch} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{"logLik"}.
}
\description{
Log-likelihood for a qgarch model
}
