% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.qgarch}
\alias{print.qgarch}
\title{Print a qgarch model}
\usage{
\method{print}{qgarch}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{A fitted \code{qgarch} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a qgarch model
}
