% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.qgarch}
\alias{print.summary.qgarch}
\title{Print a qgarch summary}
\usage{
\method{print}{summary.qgarch}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{x}{An object of class \code{"summary.qgarch"}.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a qgarch summary
}
