% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.qgarch}
\alias{summary.qgarch}
\title{Summarize a qgarch model}
\usage{
\method{summary}{qgarch}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qgarch} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{"summary.qgarch"}.
}
\description{
Summarize a qgarch model
}
