% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_monthly}
\alias{us_monthly}
\title{Monthly U.S. data}
\format{
A data frame with monthly observations and 3 variables:
\describe{
\item{YearMon}{Month of the observation.}
\item{Mktrf}{Market excess return.}
\item{ER}{Excess return series.}
}
}
\source{
Prepared by the package author.
}
\usage{
us_monthly
}
\description{
Monthly data used in examples and tests for the qgarch package.
}
\keyword{datasets}
