% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.qgarch}
\alias{vcov.qgarch}
\title{Variance-covariance matrix for a qgarch model}
\usage{
\method{vcov}{qgarch}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qgarch} object.}

\item{...}{Unused.}
}
\value{
A variance-covariance matrix.
}
\description{
Variance-covariance matrix for a qgarch model
}
