% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.qgarch}
\alias{coef.qgarch}
\title{Extract coefficients from a qgarch model}
\usage{
\method{coef}{qgarch}(object, ...)
}
\arguments{
\item{object}{A fitted \code{qgarch} object.}

\item{...}{Additional arguments. Supports \code{type = "estimated"} or
\code{type = "full"}.}
}
\value{
A named numeric vector of coefficients.
}
\description{
Extract coefficients from a qgarch model
}
