% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.qgarch}
\alias{plot.qgarch}
\title{Plot a qgarch model}
\usage{
\method{plot}{qgarch}(x, which = c("sigma2", "standardized"), ...)
}
\arguments{
\item{x}{A fitted `qgarch` object.}

\item{which}{Which plot to show: `"sigma2"` or `"standardized"`.}

\item{...}{Additional graphical arguments passed to [graphics::plot()].}
}
\value{
The input object, invisibly.
}
\description{
Plot a qgarch model
}
